+++ appendful/default.nix
@@ -3,7 +3,7 @@
 }:
 mkDerivation {
   pname = "appendful";
-  version = "0.1.0.0";
+  version = "0.2.0.0";
   src = ./.;
   libraryHaskellDepends = [
     aeson autodocodec base containers deepseq mtl validity