+++ appendful/ChangeLog.md
@@ -1,5 +1,12 @@
 # Changelog for appendful
 
+## [0.2.0.0] - 2025-09-19
+
+### Changed
+
+* The server sync processor now loads only the diff instead of the entire
+  database to process a sync request.
+
 ## [0.1.0.0] - 2022-06-19
 
 ### Added